+++ app/routes/components/ThemeProvider/changelog.codegen.ts
@@ -2,4 +2,5 @@ export const changelog = ''
   + '- 2025-09-05: Initial version.\n'
   + '- 2025-09-09: Extend DEFAULT_THEME with additional color palettes for red,\n'
   + '  green, and yellow.\n'
+  + '- 2025-09-20: Add useTheme hook and refactor ThemeProvider to make it cleaner.\n'
   ;