+++ src/components/ThemeProvider/CHANGELOG.md
@@ -1,3 +1,4 @@
 - 2025-09-05: Initial version.
 - 2025-09-09: Extend DEFAULT_THEME with additional color palettes for red,
   green, and yellow.
+- 2025-09-20: Add useTheme hook and refactor ThemeProvider to make it cleaner.